backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymupdf==1.23.8
pdf2image==1.16.3
pillow==10.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
boto3==1.34.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
openai==1.6.1
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
accelerate==0.25.0
langchain==0.0.354
langchain-community==0.0.10
pydantic-extra-types==2.4.0
stripe==7.7.0
celery==5.3.4
flower==2.0.1